requirements.txt
pyrogram@https://github.com/KurimuzonAkuma/pyrogram/archive/refs/heads/dev.zip
TgCrypto
python-i18n[YAML]